requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   google-genai
    #   httpx
attrs==25.4.0
    # via
    #   outcome
    #   trio
beautifulsoup4==4.14.3
    # via resume-builder
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
    #   selenium
cffi==2.0.0 ; (implementation_name != 'pypy' and os_name == 'nt') or platform_python_implementation != 'PyPy'
    # via
    #   cryptography
    #   trio
charset-normalizer==3.4.5
    # via
    #   pdfminer-six
    #   requests
colorama==0.4.6 ; sys_platform == 'win32'
    # via tqdm
cryptography==46.0.5
    # via
    #   google-auth
    #   pdfminer-six
distro==1.9.0
    # via google-genai
fpdf==1.7.2
    # via resume-builder
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.25.2
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.192.0
    # via google-generativeai
google-auth==2.49.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-genai
    #   google-generativeai
google-auth-httplib2==0.3.0
    # via google-api-python-client
google-genai==1.67.0
    # via resume-builder
google-generativeai==0.8.6
    # via resume-builder
googleapis-common-protos==1.73.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.78.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   wsproto
httpcore==1.0.9
    # via httpx
httplib2==0.31.2
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via google-genai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   trio
numpy==2.4.3
    # via pandas
outcome==1.3.0.post0
    # via
    #   trio
    #   trio-websocket
packaging==26.0
    # via webdriver-manager
pandas==3.0.1
    # via resume-builder
pdfminer-six==20251230
    # via pdfplumber
pdfplumber==0.11.9
    # via resume-builder
pillow==12.1.1
    # via pdfplumber
proto-plus==1.27.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.6
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.2
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0 ; (implementation_name != 'PyPy' and implementation_name != 'pypy' and os_name == 'nt') or (implementation_name != 'PyPy' and platform_python_implementation != 'PyPy')
    # via cffi
pydantic==2.12.5
    # via
    #   google-genai
    #   google-generativeai
pydantic-core==2.41.5
    # via pydantic
pyparsing==3.3.2
    # via httplib2
pypdfium2==5.6.0
    # via pdfplumber
pysocks==1.7.1
    # via urllib3
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.2
    # via webdriver-manager
requests==2.32.5
    # via
    #   google-api-core
    #   google-auth
    #   google-genai
    #   webdriver-manager
selenium==4.41.0
    # via resume-builder
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   google-genai
    #   trio
sortedcontainers==2.4.0
    # via trio
soupsieve==2.8.3
    # via beautifulsoup4
tenacity==9.1.4
    # via google-genai
tqdm==4.67.3
    # via google-generativeai
trio==0.33.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.12.2
    # via selenium
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   google-genai
    #   google-generativeai
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   selenium
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3 ; sys_platform == 'emscripten' or sys_platform == 'win32'
    # via pandas
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.6.3
    # via
    #   requests
    #   selenium
webdriver-manager==4.0.2
    # via resume-builder
websocket-client==1.9.0
    # via selenium
websockets==16.0
    # via google-genai
wsproto==1.3.2
    # via trio-websocket
